requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
spacy==3.6.0
openai==1.0.0
python-dotenv==1.0.0
groq==0.5.0